+++ src/CommonAccessors.jl
@@ -2,17 +2,22 @@
 
 # NOTE this could be reduced with macros and function generation to even less code.
 
+# Data levels
+const DataLevel0 = Union{VariableDataLevel0, FactorDataLevel0}
+const DataLevel1 = Union{VariableDataLevel1, FactorDataLevel1}
+const DataLevel2 = Union{VariableDataLevel2, FactorDataLevel2}
+
 """
 $SIGNATURES
 
-Return the label for a variable or factor.
+Return the label for a DFGNode.
 """
 getLabel(v::DataLevel0) = v.label
 
 """
 $SIGNATURES
 
-Return the tags for a variable.
+Return the tags for a DFGNode.
 """
 getTags(v::DataLevel0) = v.tags
 
@@ -20,7 +25,7 @@ getTags(v::DataLevel0) = v.tags
 """
 $SIGNATURES
 
-Set the tags for a node.
+Set the tags for a DFGNode.
 """
 function setTags!(f::DataLevel0, tags::Union{Vector{Symbol},Set{Symbol}})
   empty!(f.tags)
@@ -30,41 +35,102 @@ end
 """
 $SIGNATURES
 
-Get the timestamp from a DFGFactor object.
+Get the timestamp of a DFGNode.
 """
 getTimestamp(v::DataLevel1) = v.timestamp
 
-"""
-$SIGNATURES
-
-Set the timestamp of a DFGVariable object returning a new DFGVariable.
-Note:
-Since `timestamp` is not mutable `setTimestamp` returns a new variable with the updated timestamp.
-Use `updateVariable!` to update it in the factor graph.
-"""
-function setTimestamp(v::DFGVariable, ts::DateTime)
-    return DFGVariable(v.label, ts, v.tags, v.ppeDict, v.solverDataDict, v.smallData, v.bigData, v._dfgNodeParams)
-end
-
-function setTimestamp(v::DFGVariableSummary, ts::DateTime)
-    return DFGVariableSummary(v.label, ts, v.tags, v.estimateDict, v.softtypename, v.bigData, v._internalId)
-end
-
-setTimestamp!(f::FactorDataLevel1, ts::DateTime) = f.timestamp = ts
-
-function setTimestamp(f::DFGFactor, ts::DateTime)
-    return DFGFactor(f.label, ts, f.tags, f.solverData, f.solvable, f._dfgNodeParams, f._variableOrderSymbols)
-end
-
-function setTimestamp(f::DFGFactorSummary, ts::DateTime)
-    return DFGFactorSummary(f.label, ts, f.tags, f._internalId, f._variableOrderSymbols)
-end
 
 """
 $SIGNATURES
 
-Return the internal ID for a variable.
+Return the internal ID for DFGNode.
 """
 getInternalId(v::DataLevel2) = v._dfgNodeParams._internalId
 
 getInternalId(v::Union{DFGVariableSummary, DFGFactorSummary}) = v._internalId
+
+
+##
+
+"""
+    $SIGNATURES
+
+Variables or factors may or may not be 'solvable', depending on a user definition.  Useful for ensuring atomic transactions.
+
+Related:
+- isSolveInProgress
+"""
+getSolvable(var::Union{DFGVariable, DFGFactor})::Int = var._dfgNodeParams.solvable
+#TODO DataLevel2
+
+"""
+    $SIGNATURES
+
+Get 'solvable' parameter for either a variable or factor.
+"""
+function getSolvable(dfg::AbstractDFG, sym::Symbol)
+  if isVariable(dfg, sym)
+    return getVariable(dfg, sym)._dfgNodeParams.solvable
+  elseif isFactor(dfg, sym)
+    return getFactor(dfg, sym)._dfgNodeParams.solvable
+  end
+end
+
+
+"""
+    $SIGNATURES
+
+Variables or factors may or may not be 'solvable', depending on a user definition.
+returns true if `getSolvable` > 0
+Related:
+- `getSolvable`(@ref)
+"""
+isSolvable(node::Union{DFGVariable, DFGFactor}) = getSolvable(node) > 0
+
+
+"""
+    $SIGNATURES
+
+Which variables or factors are currently being used by an active solver.  Useful for ensuring atomic transactions.
+
+DevNotes:
+- Will be renamed to `data.solveinprogress` which will be in VND, not DFGNode -- see DFG #201
+
+Related
+
+isSolvable
+"""
+function getSolveInProgress(var::Union{DFGVariable, DFGFactor}, solveKey::Symbol=:default)::Int
+    # Variable
+    var isa DFGVariable && return haskey(getSolverDataDict(var), solveKey) ? getSolverDataDict(var)[solveKey].solveInProgress : 0
+    # Factor
+    return getSolverData(var).solveInProgress
+end
+
+
+isSolveInProgress(node::Union{DFGVariable, DFGFactor}, solvekey::Symbol=:default) = getSolveInProgress(node, solvekey) > 0
+
+
+"""
+    $SIGNATURES
+
+Set the `solvable` parameter for either a variable or factor.
+"""
+function setSolvable!(dfg::AbstractDFG, sym::Symbol, solvable::Int)::Int
+  if isVariable(dfg, sym)
+    getVariable(dfg, sym)._dfgNodeParams.solvable = solvable
+  elseif isFactor(dfg, sym)
+    getFactor(dfg, sym)._dfgNodeParams.solvable = solvable
+  end
+  return solvable
+end
+
+"""
+    $SIGNATURES
+
+Set the `solvable` parameter for either a variable or factor.
+"""
+function setSolvable!(node::N, solvable::Int)::Int where N <: DFGNode
+  node._dfgNodeParams.solvable = solvable
+  return solvable
+end
