+++ src/entities/DFGVariable.jl
@@ -1,4 +1,13 @@
-## VariableNodeData.jl
+##==============================================================================
+## Abstract Types
+##==============================================================================
+
+abstract type InferenceVariable end
+
+##==============================================================================
+## VariableNodeData
+##==============================================================================
+
 """
 $(TYPEDEF)
 Data container for solver-specific data.
@@ -45,6 +54,9 @@ mutable struct VariableNodeData{T<:InferenceVariable}
                                    dontmargin, solveInProgress, solvedCount)
 end
 
+##------------------------------------------------------------------------------
+## Constructors
+
 VariableNodeData(val::Array{Float64,2},
                  bw::Array{Float64,2},
                  BayesNetOutVertIDs::Array{Symbol,1},
@@ -63,11 +75,15 @@ VariableNodeData(val::Array{Float64,2},
                                        eliminated,BayesNetVertID,separator,
                                        softtype::T,initialized,inferdim,ismargin,
                                        dontmargin, solveInProgress, solvedCount)
-#
+
+
 VariableNodeData(softtype::T) where T <: InferenceVariable =
     VariableNodeData{T}(zeros(1,1), zeros(1,1), Symbol[], Int[], 0, false, :NOTHING, Symbol[], softtype, false, 0.0, false, false, 0, 0)
 
-##PackedVariableNodeData.jl
+##==============================================================================
+## PackedVariableNodeData.jl
+##==============================================================================
+
 """
 $(TYPEDEF)
 Packed VariabeNodeData structure for serializing DFGVariables.
@@ -114,9 +130,19 @@ mutable struct PackedVariableNodeData
                          solvedCount::Int) = new(x1,x2,x3,x4,x5,x6,x7,x8,x9,x10,x11,x12,x13,x14,x15,x16, solvedCount)
 end
 
-## PoiintParametricEst.jl
-# AbstractPointParametricEst interface
+##==============================================================================
+## PointParametricEst
+##==============================================================================
+
+##------------------------------------------------------------------------------
+## AbstractPointParametricEst interface
+##------------------------------------------------------------------------------
+
 abstract type AbstractPointParametricEst end
+
+##------------------------------------------------------------------------------
+## MeanMaxPPE
+##------------------------------------------------------------------------------
 """
     $TYPEDEF
 
@@ -129,14 +155,26 @@ struct MeanMaxPPE <: AbstractPointParametricEst
     mean::Vector{Float64}
     lastUpdatedTimestamp::DateTime
 end
+##------------------------------------------------------------------------------
+## Constructors
+
 MeanMaxPPE(solverKey::Symbol, suggested::Vector{Float64}, max::Vector{Float64},mean::Vector{Float64}) = MeanMaxPPE(solverKey, suggested, max, mean, now())
+
+## Accessors
+##------------------------------------------------------------------------------
 
 getMaxPPE(est::AbstractPointParametricEst) = est.max
 getMeanPPE(est::AbstractPointParametricEst) = est.mean
 getSuggestedPPE(est::AbstractPointParametricEst) = est.suggested
 getLastUpdatedTimestamp(est::AbstractPointParametricEst) = est.lastUpdatedTimestamp
 
-## DFGVariable.jl
+##==============================================================================
+## DFG Variables
+##==============================================================================
+
+##------------------------------------------------------------------------------
+## DFGVariable lv2
+##------------------------------------------------------------------------------
 """
 $(TYPEDEF)
 Complete variable structure for a DistributedFactorGraph variable.
@@ -162,7 +200,7 @@ struct DFGVariable{T<:InferenceVariable} <: AbstractDFGVariable
     Accessors: `addVariableSolverData!`, `updateVariableSolverData!`, and `deleteVariableSolverData!`"""
     solverDataDict::Dict{Symbol, VariableNodeData{T}}
     """Dictionary of small data associated with this variable.
-    Accessors: `addSmallData!`, `updateSmallData!`, and `deleteSmallData!`"""
+    Accessors: [`getSmallData!`](@ref), [`setSmallData!`](@ref)"""
     smallData::Dict{String, String}#Ref{Dict{String, String}} #why was Ref here?
     """Dictionary of large data associated with this variable.
     Accessors: `addBigDataEntry!`, `getBigDataEntry`, `updateBigDataEntry!`, and `deleteBigDataEntry!`"""
@@ -171,6 +209,9 @@ struct DFGVariable{T<:InferenceVariable} <: AbstractDFGVariable
     Accessors: `getSolvable`, `setSolvable!`"""
     _dfgNodeParams::DFGNodeParams
 end
+
+##------------------------------------------------------------------------------
+## Constructors
 
 """
     $SIGNATURES
@@ -199,13 +240,17 @@ DFGVariable(label::Symbol,
             _internalId::Int64=0) where {T <: InferenceVariable} =
     DFGVariable{T}(label, timestamp, tags, estimateDict, Dict{Symbol, VariableNodeData{T}}(:default=>solverData), smallData, bigData, DFGNodeParams(solvable, _internalId))
 
+##------------------------------------------------------------------------------
 function Base.copy(o::DFGVariable)::DFGVariable
     return DFGVariable(o.label, getSofttype(o)(), tags=copy(o.tags), estimateDict=copy(o.estimateDict),
                         solverDataDict=copy(o.solverDataDict), smallData=copy(o.smallData),
                         bigData=copy(o.bigData), solvable=getSolvable(o), _internalId=getInternalId(o))
 end
 
-## DFGVariableSummary.jl
+##------------------------------------------------------------------------------
+## DFGVariableSummary lv1
+##------------------------------------------------------------------------------
+
 """
 $(TYPEDEF)
 Summary variable structure for a DistributedFactorGraph variable.
@@ -237,7 +282,10 @@ struct DFGVariableSummary <: AbstractDFGVariable
     _internalId::Int64
 end
 
+##------------------------------------------------------------------------------
 ## SkeletonDFGVariable.jl
+##------------------------------------------------------------------------------
+
 """
 $(TYPEDEF)
 Skeleton variable structure for a DistributedFactorGraph variable.
@@ -256,3 +304,223 @@ struct SkeletonDFGVariable <: AbstractDFGVariable
 end
 
 SkeletonDFGVariable(label::Symbol) = SkeletonDFGVariable(label, Set{Symbol}())
+
+
+##==============================================================================
+# Define variable levels
+##==============================================================================
+const VariableDataLevel0 = Union{DFGVariable, DFGVariableSummary, SkeletonDFGVariable}
+const VariableDataLevel1 = Union{DFGVariable, DFGVariableSummary}
+const VariableDataLevel2 = Union{DFGVariable}
+
+
+
+function Base.convert(::Type{DFGVariableSummary}, v::DFGVariable)
+    return DFGVariableSummary(v.label, v.timestamp, deepcopy(v.tags), deepcopy(v.ppeDict), Symbol(typeof(getSofttype(v))), v.bigData, v._internalId)
+end
+
+#TODO Test
+function Base.convert(::Type{SkeletonDFGVariable}, v::VariableDataLevel1)
+    return SkeletonDFGVariable(v.label, deepcopy(v.tags))
+end
+
+
+##==============================================================================
+## Accessors
+##==============================================================================
+
+"""
+$SIGNATURES
+
+Set the timestamp of a DFGVariable object returning a new DFGVariable.
+Note:
+Since `timestamp` is not mutable `setTimestamp` returns a new variable with the updated timestamp.
+Use `updateVariable!` to update it in the factor graph.
+"""
+function setTimestamp(v::DFGVariable, ts::DateTime)
+    return DFGVariable(v.label, ts, v.tags, v.ppeDict, v.solverDataDict, v.smallData, v.bigData, v._dfgNodeParams)
+end
+
+function setTimestamp(v::DFGVariableSummary, ts::DateTime)
+    return DFGVariableSummary(v.label, ts, v.tags, v.estimateDict, v.softtypename, v.bigData, v._internalId)
+end
+
+
+"""
+   $(SIGNATURES)
+
+Variable nodes softtype information holding a variety of meta data associated with the type of variable stored in that node of the factor graph.
+
+Related
+
+getVariableType
+"""
+function getSofttype(vnd::VariableNodeData)
+  return vnd.softtype
+end
+function getSofttype(v::DFGVariable, solvekey::Symbol=:default)
+  return v.solverDataDict[solvekey].softtype # Get instantiated form of the parameter for the DFGVariable
+end
+
+"""
+    $SIGNATURES
+
+Retrieve the soft type name symbol for a DFGVariableSummary. ie :Point2, Pose2, etc.
+TODO, DO NOT USE v.softtypename in DFGVariableSummary
+"""
+getSofttype(v::DFGVariableSummary)::Symbol = v.softtypename
+
+
+"""
+    $SIGNATURES
+
+Retrieve solver data structure stored in a variable.
+"""
+function getSolverData(v::DFGVariable, key::Symbol=:default)
+    #TODO this does not fit in with some of the other error behaviour. but its used so added @error
+    vnd =  haskey(v.solverDataDict, key) ? v.solverDataDict[key] : (@error "Variable does not have solver data $(key)"; nothing)
+    return vnd
+end
+
+"""
+    $SIGNATURES
+
+Set solver data structure stored in a variable.
+"""
+#TODO Repeated functionality?
+setSolverData!(v::DFGVariable, data::VariableNodeData, key::Symbol=:default) = v.solverDataDict[key] = data
+
+"""
+    $SIGNATURES
+
+Get solver data dictionary for a variable.
+"""
+getSolverDataDict(v::DFGVariable) = v.solverDataDict
+
+"""
+    $SIGNATURES
+
+Get the PPE dictionary for a variable. Its use is not recomended.
+"""
+getPPEDict(v::VariableDataLevel1) = v.ppeDict
+
+
+#TODO FIXME don't know if this should exist, should rather always update with fg object to simplify inmem vs cloud
+"""
+    $SIGNATURES
+
+Get the parametric point estimate (PPE) for a variable in the factor graph.
+
+Notes
+- Defaults on keywords `solveKey` and `method`
+
+Related
+
+getMeanPPE, getMaxPPE, getKDEMean, getKDEFit, getPPEs, getVariablePPEs
+"""
+function getPPE(vari::VariableDataLevel1, solveKey::Symbol=:default)
+    return  getPPEs(vari)[solveKey]
+    # return haskey(ppeDict, solveKey) ? ppeDict[solveKey] : nothing
+end
+
+
+
+"""
+$SIGNATURES
+
+Get the small data for a variable.
+"""
+getSmallData(v::DFGVariable)::Dict{String, String} = v.smallData
+
+"""
+$SIGNATURES
+
+Set the small data for a variable.
+This will overwrite old smallData.
+"""
+function setSmallData!(v::DFGVariable, smallData::Dict{String, String})::Dict{String, String}
+    empty!(v.smallData)
+    merge!(v.smallData, smallData)
+end
+
+
+"""
+$SIGNATURES
+
+Get the variable ordering for this factor.
+Should be equivalent to getNeighbors unless something was deleted in the graph.
+"""
+getVariableOrder(fct::DFGFactor)::Vector{Symbol} = fct._variableOrderSymbols
+getVariableOrder(dfg::AbstractDFG, fct::Symbol) = getVariableOrder(getFactor(dfg, fct))
+
+"""
+    $SIGNATURES
+
+Get the number of times a variable has been inferred -- i.e. `solvedCount`.
+
+Related
+
+isSolved, setSolvedCount!
+"""
+getSolvedCount(v::VariableNodeData) = v.solvedCount
+getSolvedCount(v::VariableDataLevel2, solveKey::Symbol=:default) = getSolverData(v, solveKey) |> getSolvedCount
+getSolvedCount(dfg::AbstractDFG, sym::Symbol, solveKey::Symbol=:default) = getSolvedCount(getVariable(dfg, sym), solveKey)
+
+"""
+    $SIGNATURES
+
+Boolean on whether the variable has been solved.
+
+Related
+
+getSolved, setSolved!
+"""
+isSolved(v::VariableNodeData) = 0 < v.solvedCount
+isSolved(v::VariableDataLevel2, solveKey::Symbol=:default) = getSolverData(v, solveKey) |> isSolved
+isSolved(dfg::AbstractDFG, sym::Symbol, solveKey::Symbol=:default) = isSolved(getVariable(dfg, sym), solveKey)
+
+
+"""
+    $SIGNATURES
+
+Update/set the `solveCount` value.
+
+Related
+
+getSolved, isSolved
+"""
+setSolvedCount!(v::VariableNodeData, val::Int) = v.solvedCount = val
+setSolvedCount!(v::VariableDataLevel2, val::Int, solveKey::Symbol=:default) = setSolvedCount!(getSolverData(v, solveKey), val)
+setSolvedCount!(dfg::AbstractDFG, sym::Symbol, val::Int, solveKey::Symbol=:default) = setSolvedCount!(getVariable(dfg, sym), val, solveKey)
+
+"""
+    $SIGNATURES
+
+Returns state of vertex data `.initialized` flag.
+
+Notes:
+- used by both factor graph variable and Bayes tree clique logic.
+"""
+function isInitialized(var::DFGVariable, key::Symbol=:default)::Bool
+      data = getSolverData(var, key)
+      if data == nothing
+          #TODO we still have a mixture of 2 error behaviours
+        return false
+      else
+          return data.initialized
+    end
+end
+
+function isInitialized(dfg::AbstractDFG, label::Symbol, key::Symbol=:default)::Bool
+  return isInitialized(getVariable(dfg, label), key)
+end
+
+"""
+    $SIGNATURES
+
+Return the DFGVariable softtype in factor graph `dfg<:AbstractDFG` and label `::Symbol`.
+"""
+getVariableType(var::DFGVariable) = getSofttype(var)
+function getVariableType(dfg::G, lbl::Symbol) where G <: AbstractDFG
+  getVariableType(getVariable(dfg, lbl))
+end
